### 2016-08-16 #51 @whedon
Submission: rdefra: Interact with the UK AIR Pollution Database from DEFRA
**Submitting author:** @cvitolo (<a href="http://orcid.org/0000-0002-4252-1176">Claudia Vitolo</a>)
**Repository:** <a href="https://github.com/kehraProject/r_rdefra" target ="_blank">https://github.com/kehraProject/r_rdefra</a>
**Version:** v0.3.0
