### 2016-08-19 #51 @pjotrp
Web scraping is a notoriously fragile method.

This software consists of 243 lines of R code. I would like to ask the
author to confirm *here* that he thinks this software is a valuable
contribution to research. I want to point out that once this work has
been published it will be visible for a long time. Adding an example
of having used this tool in research would be a good idea.
