### 2016-08-21 #51 @pragyansmita
This is part of the openSci R package and provides access to the Air Information Resource UK-AIR of the Department for Environment, Food and Rural Affairs in the United Kingdom (no public API available today).

@cvitolo - Can you please confirm the following are included?

-  a plain-text LICENSE file with the contents of an OSI approved software license
- release version number in documentation matching Github

The readme references in usage ukair_catalogue(). However, that method is no longer available. Using catalogue() is available instead. ukair_get_coordinates is also not available. Can the readme and provided examples please be updated to match the current methods?

If any automated tests can be provided apart from the provided examples, please include the same.
