### 2016-08-26 #51 @arfon
> a plain-text LICENSE file with the contents of an OSI approved software license

@pragyansmita - CRAN has some pretty restrictive policies around where the license can be present. See this thread for background: https://github.com/openjournals/joss-reviews/issues/37#issuecomment-231787026

I believe that @cvitolo has adhered to the CRAN guidelines here which is acceptable to JOSS.
