### 2016-08-28 #51 @cvitolo
Many thanks @arfon - I have just made a new release v0.3.2, here is the doi url: http://dx.doi.org/10.5281/zenodo.61033
