### 2016-08-28 #51 @arfon
Thanks for the review @pragyansmita :zap:

@cvitolo - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00051 🎉 🚀 💥
