### 2016-09-13 #51 @noamross
Adding a note for reference: *rdefra* went through JOSS and rOpenSci review in parallel, having been submitted before our joint-review workflow was in place but approved after.  The rOpenSci review can be found here: https://github.com/ropensci/onboarding/issues/68
